Fix compile warnings for MinGW on Debian Stretch

+++ yaesu/ft890.c
@@ -1906,8 +1906,8 @@ static int ft890_send_rit_freq(RIG *rig, unsigned char ci, shortfreq_t rit)
     to_bcd(priv->p_cmd, rit / 10, FT890_BCD_RIT);
 
     rig_debug(RIG_DEBUG_TRACE,
-              "%s: requested rit after conversion = %li Hz\n",
-              __func__, (int64_t)from_bcd(priv->p_cmd, FT890_BCD_RIT) * 10);
+              "%s: requested rit after conversion = %d Hz\n",
+              __func__, (int)from_bcd(priv->p_cmd, FT890_BCD_RIT) * 10);
 
     priv->p_cmd[P1] = p1;         /* ick */
     priv->p_cmd[P2] = p2;
